requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Validation
pydantic[email]>=2.7.0
pydantic-settings>=2.3.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.9

# HTTP client (for external APIs — Railway, GitHub, etc.)
httpx==0.27.0

# AI
anthropic>=0.40.0

# Cache / Queue
redis[hiredis]==5.0.2

# Env
python-dotenv==1.0.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0  # also used as async test client
aiosqlite==0.20.0  # async SQLite driver for in-memory test DB
